Properly support the 'respawn_min_delay' configuration option
This support was damaged during the transition to the internal pipe. This
commit fixes this option in a much cleaner manner. Before poll()ing the
internal pipe, sl2d computes the precise time it needs to sleep before
launching the next process.

FTR, this is how the glib main loop works when using g_timeout_add().